refactor(core/rust): move InputStream to its own module
