+++ core/embed/rust/src/lib.rs
@@ -17,6 +17,8 @@ mod error;
 // use trezorhal for its macros early
 #[macro_use]
 mod trezorhal;
+
+mod io;
 mod maybe_trace;
 #[cfg(feature = "micropython")]
 #[macro_use]
